♻️ Extract cookie modules into helpers/server/cookie/
- cookie/types.ts: CookieToSet type
- cookie/helpers.ts: getCookieOptions() with domain
- cookie/auth.cookie.ts: SESSION_COOKIE, REFRESH_COOKIE, buildSessionCookies()
- cookie/region.cookie.ts: REGION_COOKIE, buildRegionCookie() (moved from region.middleware)

Removes proxy/types.ts dependency on inline CookieToSet type.
Cleans up region.middleware.ts — no longer hosts cookie constants.
