+++ apps/site/src/helpers/server/cookie/auth.cookie.ts
@@ -1,31 +1,10 @@
 import { REFRESH_TOKEN_TTL_DAYS } from '@nosgestesclimat/core/features/auth/constants/session'
 import type { SessionTokens } from '@nosgestesclimat/core/features/auth/types/session'
-import type { NextResponse } from 'next/server'
-
-type CookieOptions = NonNullable<Parameters<NextResponse['cookies']['set']>[2]>
-
-export interface CookieToSet {
-  name: string
-  value: string
-  options?: Partial<CookieOptions>
-}
-
-const domain = new URL(process.env.NEXT_PUBLIC_SITE_URL!).hostname
-const secure = domain !== 'localhost'
+import { getCookieOptions } from './helpers'
+import type { CookieToSet } from './types'
 
 export const SESSION_COOKIE = 'ngc_session'
 export const REFRESH_COOKIE = 'ngc_refresh'
-
-export function getCookieOptions() {
-  return {
-    httpOnly: true,
-    secure,
-    sameSite: secure ? ('none' as const) : ('lax' as const),
-    partitioned: secure,
-    path: '/',
-    ...(secure ? { domain } : {}),
-  }
-}
 
 export function buildSessionCookies(tokens: SessionTokens): CookieToSet[] {
   const maxAge = REFRESH_TOKEN_TTL_DAYS * 24 * 60 * 60
