+++ apps/site/src/helpers/server/cookie/helpers.ts
@@ -1,0 +1,13 @@
+const domain = new URL(process.env.NEXT_PUBLIC_SITE_URL!).hostname
+const secure = domain !== 'localhost'
+
+export function getCookieOptions() {
+  return {
+    httpOnly: true,
+    secure,
+    sameSite: secure ? ('none' as const) : ('lax' as const),
+    partitioned: secure,
+    path: '/',
+    ...(secure ? { domain } : {}),
+  }
+}
