+++ apps/site/src/helpers/server/cookie/types.ts
@@ -1,0 +1,9 @@
+import type { NextResponse } from 'next/server'
+
+type CookieOptions = NonNullable<Parameters<NextResponse['cookies']['set']>[2]>
+
+export interface CookieToSet {
+  name: string
+  value: string
+  options?: Partial<CookieOptions>
+}
