+++ apps/site/src/helpers/server/proxy/types.ts
@@ -1,14 +1,5 @@
+import type { CookieToSet } from '@/helpers/server/cookie/types'
 import type { NextResponse } from 'next/server'
-
-type CookieOptions = NonNullable<
-  Parameters<NextResponse['cookies']['set']>[2]
->
-
-export interface CookieToSet {
-  name: string
-  value: string
-  options?: Partial<CookieOptions>
-}
 
 export interface MiddlewareResult {
   redirect: NextResponse | null
